// secure-agents/frontend/src/client/types.gen.ts
export type UserPublic = {
  /**
   * Id
   */
  id: string;
  /**
   * Email
   */
  email?: string | null;
  /**
   * Full Name
   */
  full_name?: string | null;
  /**
   * Preferred Username
   */
  preferred_username?: string | null;
  /**
   * Unique Name
   */
  unique_name?: string | null;
  /**
   * Upn
   */
  upn?: string | null;
};